refactor: rename GetStartedCard to HistoryCard and remove unused css
