--- src/components/GetStartedCard/GetStartedCard.module.css
@@ -1,7 +1,0 @@
-.header {
-  margin-bottom: 0.5em;
-}
-
-.heading {
-  margin-right: 0.5em;
-}
