+++ src/components/MainContent/index.tsx
@@ -1,11 +1,11 @@
 import { PropsWithChildren } from 'react'
-import GetStartedCard from '../GetStartedCard'
+import HistoryCard from '../HistoryCard'
 import styles from './MainContent.module.css'
 
 export default function MainContent(props: PropsWithChildren) {
   return (
     <div className={styles.mainContent}>
-      <GetStartedCard />
+      <HistoryCard />
     </div>
   )
 }
